# backend/nest/apiswagger.yaml
openapi: 3.0.1
info:
  title: FStudy - Language Course Purchase API
  description: |
    Welcome to the FStudy API! This API allows users to browse, purchase, and manage language courses on the FStudy platform. Our platform offers a wide range of language courses designed to help learners achieve their language learning goals efficiently and effectively.

    ## Features
    - **Course Catalog**: Browse a comprehensive list of available language courses.
    - **Course Details**: Get detailed information about each course, including syllabus, instructor information, and pricing.
    - **User Management**: Manage user profiles, including registration, login, and profile updates.
    - **Purchase Courses**: Securely purchase language courses and access purchased content.
    - **Progress Tracking**: Track learning progress and achievements within the platform.
    - **Reviews and Ratings**: Read and submit reviews and ratings for courses.

    ## Authentication
    The FStudy API uses JWT (JSON Web Token) for authentication. Users must obtain a token by logging in and include it in the `Authorization` header of requests.

    ## Error Handling
    The API returns standard HTTP status codes for successful and erroneous requests. Error responses include a message explaining the reason for the error.

    ## Base URL
    The base URL for all API requests is: `http://localhost:4000/`

    ## Versioning
...

    ## Contact
...

  contact:
    name: FStudy Support
...
  license:
...
servers:
...
paths:
...
  /blog/create-topic:
    post:
      tags:
      - Blogs
      summary: Create a new topic
      description: This endpoint allows you to create a new blog topic.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the topic
                  example: Toeic 123
                description:
                  type: string
                  description: Description of the topic
                  example: Đây là Toeic
                subTopics:
                  type: string
                  description: IDs of subtopics (optional)
                  example: 66977643badef7824986a2a6
                  nullable: true
      responses:
        '200':
          description: Topic created successfully
  /blog/update-topic/{id}:
    put:
      tags:
      - Blogs
      summary: Update a topic
      description: This endpoint allows you to update an existing blog topic.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 669c70b6b8a925b598c50ce2
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the topic
                  example: Ielt 123 456
                description:
                  type: string
                  description: Description of the topic
                  example: This is Ielts's blogs
                subTopics:
                  type: string
                  description: IDs of subtopics (optional)
                  example: 66977643badef7824986a2a6
                  nullable: true
      responses:
        '200':
          description: Topic updated successfully
  /blog/topic/{id}:
    get:
      tags:
      - Blogs
      summary: Get topic by ID
      description: This endpoint allows you to retrieve a blog topic by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 6697b7975fbb6c1462fb4e3c
      responses:
        '200':
          description: Topic retrieved successfully
  /blog/topic/{iid}:
    delete:
      tags:
      - Blogs
      summary: Delete topic by ID
      description: This endpoint allows you to delete a blog topic by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 6697b7f95fbb6c1462fb4e43
      responses:
        '200':
          description: Topic deleted successfully
...
  /blog/create-child-topic:
    post:
      tags:
      - Blogs
      summary: Create a new child topic
      description: This endpoint allows you to create a new child topic under a specific blog topic.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the child topic
                  example: Toeic moi nhat 2024
                description:
                  type: string
                  description: Description of the child topic
                  example: Đây là Toeic 1
                topic:
                  type: string
                  description: ID of the parent topic
                  example: 6697b7f95fbb6c1462fb4e43
      responses:
        '200':
          description: Child topic created successfully
  /blog/update-child-topic/{id}:
    put:
      tags:
      - Blogs
      summary: Update a child topic
      description: This endpoint allows you to update an existing child topic.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 669c70e8b8a925b598c50ce5
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the child topic
                  example: Speaking moi nhat 2025
                description:
                  type: string
                  description: Description of the child topic
                  example: This is Speaking part of blog
                topic:
                  type: string
                  description: ID of the parent topic
                  example: 6697b7f95fbb6c1462fb4e43
      responses:
        '200':
          description: Child topic updated successfully
  /blog/child-topic/{id}:
    get:
      tags:
      - Blogs
      summary: Get child topic by ID
      description: This endpoint allows you to retrieve a child topic by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 6697b8f4a25097ac65745d3b
      responses:
        '200':
          description: Child topic retrieved successfully
  /blog/child-topic/{iid}:
    delete:
      tags:
      - Blogs
      summary: Delete child topic by ID
      description: This endpoint allows you to delete a child topic by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 669c70e8b8a925b598c50ce5
      responses:
        '200':
          description: Child topic deleted successfully
...
  /blog/search:
    post:
      tags:
      - Blogs
      summary: Search blogs
      description: This endpoint allows you to search blogs by a keyword.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                key:
                  type: string
                  description: The search keyword
                  example: Bài viết 1
      responses:
        '200':
          description: Blogs retrieved successfully
  /blog/update-blog/{id}:
    put:
      tags:
      - Blogs
      summary: Update a blog by ID
      description: This endpoint allows you to update an existing blog by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 669c7130b8a925b598c50ce8
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the blog
                  example: Bài viết 1
                content:
                  type: string
                  description: Content of the blog
                  example: đây là content của bài viết 1
                title:
                  type: string
                  description: Title of the blog
                  example: đây là title của bài viết 1 moi dươc sửa nè
                avatar:
                  type: string
                  format: binary
                  description: Avatar image of the blog
                status:
                  type: string
                  description: Status of the blog
                  example: 1
                childTopics:
                  type: string
                  description: ID of the child topics
                  example: 6697bbb1e6302e630069c8a7
      responses:
        '200':
          description: Blog updated successfully
  /blog/create-blog:
    post:
      tags:
      - Blogs
      summary: Create a new blog
      description: This endpoint allows you to create a new blog.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the blog
                  example: đây là bài viết mới được viết luôn nè
                content:
                  type: string
                  description: Content of the blog
                  example: đây là content của bài viết 2
                title:
                  type: string
                  description: Title of the blog
                  example: đây là title của bài viết 232
                avatar:
                  type: string
                  format: binary
                  description: Avatar image of the blog
                status:
                  type: string
                  description: Status of the blog
                  example: 1
                childTopics:
                  type: string
                  description: ID of the child topics
                  example: 6697bbb1e6302e630069c8a7
      responses:
        '200':
          description: Blog created successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer 
      bearerFormat: JWT